We should not mount /dev read-only
/dev is on a tmpfs.  It is not part of the host file system, so it should not be read-only.

This stops certain containers from creating device nodes in /dev as well as /dev/log.

Breaks systemd, which wants to create /dev/log for the journal.  Also will also break any
tool that creates a tty device or any of the other devices allowed to be created by the device cgroup.

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>