Create the xattr test file in user's home dir
Running tests on a Mac host with a VirtualBox Linux docker host we see
the TestXattr fail because the test is unable to set the extended
attributes on /go/src/github.com/opencontainers/runc. Extended
attributes do not seem to work on VBox file systems (which is what
VirtualBox mounts the runc directory as).

We propose creating the test file in the user's home directory which is
more likely to be a file system that supports extended attributes

We think this change is uncontroversial because of this statement in the
README:
https://github.com/opencontainers/runc/blame/10dc96bb9e7e3b07511efdc830560c1384d92f79/README.md#L72

Signed-off-by: Danail Branekov <danailster@gmail.com>