libcontainer/mount: use Getfsstat instead of C.getmntinfo on FreeBSD
The getmntinfo C function on FreeBSD uses an internal static buffer to
store the results and returns a pointer to this buffer. This is not
thread safe. Instead do the lookup using getfsstat (which getmntinfo
also does) but allocate the buffer in the caller.

This allows to get rid of cgo usage in libcontainer.

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>