libcontainer/standard_init_linux: Move LookPath to post-wait
This avoids a panic for containers that do not set Process.  And even
if Process was set, there is no reason to require the executable to be
available *at create time* [1].  Subsequent activity could be
scheduled to get a binary in place at the configured location before
'start' is called.

[1]: https://github.com/opencontainers/runc/pull/827#discussion_r63464201

Signed-off-by: W. Trevor King <wking@tremily.us>