Provide install instruction that uses "go get"
Is there any substantial difference to using go get instead of git clone? I first did the git clone (and of course) cloned one directory level off, resulting in the wrong paths and a bunch of errors. This is of course user error, but given that Go is a dependency anyway, would it make sense to control this clone step to prevent the user (me<--) from messing it  up?