kill: add an option --signal to specify a signal
According to the command line interface of runtime-tools,
https://github.com/opencontainers/runtime-tools/blob/master/docs/command-line-interface.md#kill,
signal needs to be set by an option `--signal`. However, `runc kill`
does not have the option, but just takes the first argument as the
signal.

Let's add a new option `--signal` to `runc kill`. Then we are able to
make runtime-tools call simply a kill command with `--signal` in
https://github.com/opencontainers/runtime-tools/blob/master/validation/util/container.go#L172..L174.

Signed-off-by: Dongsu Park <dongsu@kinvolk.io>