Set default CRIU WorkDirectory = ImagesDirectory
CRIU is using both images and work directory to store files during
checkpoint and restore. The work directory contains files such as logs
and irmap cache.

Currently runc sets the default value of the work directory to
`filepath.Join(c.root, "criu.work")`, and in case of failure outputs
the following message:

    # runc restore --image-path /tmp/test/ mycontainer
    criu failed: type NOTIFY errno 0
    log file: /run/runc/mycontainer/criu.work/restore.log

However, this is not very useful because the directory
`/run/runc/mycontainer/criu.work` is removed in case of failure, and
therefore the `restore.log` file does not exist.

This patch changes runc to follow the default behaviour of CRIU by
setting the working directory to be the same as the images directory,
and therefore preserve the log file in case of failure.

Signed-off-by: Radostin Stoyanov <rstoyanov1@gmail.com>