Add public API interface
Main motivation of this commit is that dependent projects do not have to
shell out for using runc. This provides better performance and a general
type-safety.

This commit adds a public API to runc, which currently implements the
basic Version() API for demonstration purposes. A public interface
provides the possibility to mock the results by the consumer of the API.
The configuration of the API will be done via a modified builder
pattern, which is directly part of the public interface.

Later on, the all CLI commands within the `main` package should utilize
the configured API instance from the context to retrieve its results.

Signed-off-by: Sascha Grunert <sgrunert@suse.com>