Split support for unified cgroup2 hierarchy into its own Manager
Detection is done only once, as the manager is instantiated, and then the two
implementations are for the most part separate.

This will allow us to iterate on the unified cgroup2 implementation without
worrying about compatibility with the legacy implementation.

For now, the unified cgroup2 implementation consists of stubs that need to be
filled up with actual implementation.

Signed-off-by: Filipe Brandenburger <filbranden@gmail.com>