nsenter: cloned_binary: do not clone binary if on readonly rootfs
Fix for CVE-2019-5736 introduces a logic to clone the /proc/self/exe
so that no malicious container can modify the host binary. Due to this
bind mount was done in try_bindfd() which introduces three libmount
events. In systemd <= 232, for every libmount event, iteration over
all mounts present in /proc/self/mountinfo is done to check which
mount got changed. In case of large number of mounts already present,
systemd will require high cpu utilization for iteration over all the
mounts. In case of a read-only rootfs, cloning of /proc/self/exe is
not required and therefore, generation of libmount events can be
avoided.

This patch adds a check to see whether /proc/self/exe is on a
read-only rootfs or not. In case, it is present on a read-only
rootfs, cloning is not required.

Signed-off-by: Vaibhav Rustagi <vaibhavrustagi@google.com>