libcontainer/stacktrace: fix inline func name
The TestCaptureTestFunc fails on gccgo, because captureFunc is inlined
there. Instead of using FuncForPC directly, using CallersFrames from
runtime package, which is available since go1.7

This commit has an API break for the stacktrace package, which removes
NewFrame func. But I think it's safe as the stacktrace is somehow an
internal package of libcontainer.

Signed-off-by: Shengjing Zhu <zhsj@debian.org>