newSetnsProcess: don't call GetPaths for v2
The setnsProcess.cgroupPaths map is used by (*setnsProcess).start
to enter the cgroups. Since in case of cgroupv2 unified hierarchy
there is only one cgroup to enter, let's create a map with a
single element in it. As key does not matter it is empty.

As a side effect, we won't write the pid into the same file 7 times :)

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>