cgroups: move default CpuPeriod to specconv
In case CpuQuota is set, and CpuPeriod is not, we should assume the
default (which is the same for v1 and v2). Currently, only
ConvertCPUQuotaCPUPeriodToCgroupV2Value assumes it.

Move the assumption to specconv, making it universally available.
This makes all the cgroup drivers behave the same.

This commit also fixes the regression introduced by commit
06d7c1d2616052974e, which made it impossible to only set CpuQuota
(without the CpuPeriod).

NOTE it is still wrong to set CpuPeriod without setting CpuQuota.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>