cgroupv1/Apply: do not overuse path/getSubsystemPath
When paths are already set, we only need to place the PID into proper
cgroups, and we already know all the cgroups path.

The fs function d.path(), and the systemd v1 function getSubsystemPath()
are both parsing /proc/self/mountinfo, and the only reason they are used
here is to check whether the subsystem is available.

Use a much simpler check instead.

Frankly, I am not sure why the check is needed at all. Maybe we just
need to drop it altogether.

Also, for fs, since d is no longer used in this code path, moves its
initialization to after it.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>